Add files via upload
Caesar Cipher
Cryptography is one of the important branches one should learn to get good at cyber security. A lot of concepts like encryption and decryption, ciphers, hashes etc. are to be learnt. Let's start with Caesar Cipher.
Caesar cipher is an olden cipher which is not in use anymore because it's not secure and pretty easy to crack.
It just substitutes every letter of the message with a letter at an offset of a number called 'key'.
Example: Let message be 'Anjum Shaik' and key is 5.
The encrypted message will be 'Fs0zr Xmfnp'
Because 'A'+ 5 letters = 'F' ….. and so on.
This is how it decrypts the messages too. It just subtracts the number of letters instead of addition.
It is pretty easy to crack as there are limited number of keys and any key exceeding the number of alphabets will be counted from the start after subtracting the number of alphabets and vice versa.